restore/finetune: allow non-strict loading of state dict
This only applies to the `finetune` mode.

Enables a new boolean config key `train.restore.strict` to be used. Its value is passed 
as the `strict` argument of the `load_state_dict(checkpoint["state_dict"])` function.
This way, the strictness of this phase can be user-definable, and optional.

The motivation being that sometimes it is useful to finetune a LightningModule
that is largely the same as the one we're loading, but with additional weights 
that were not originally present in the checkpoint.